+++ ChangeLog.md
@@ -3242,7 +3242,19 @@ Exit Criteria:
 
 Some preliminary test events will need to be created for this chapter, but a separate, content issue will be created upon the feature's completion as well.
 
-### 1. Step One...
+1. Create a new type, RandomEventData, which has a karma value as well as a magnitude value, and a data component which is an ordinary EventData type object.
+
+2. Create a new World Builder function that can add random event data of the above type to the database's random_events collection.
+
+3. Create a new server function that can fetch random events from the database and return one at random.
+
+### 4. Convert the existing random events in the frontend's constants file into RandomEventData objects in the world builder's newEvents file. For now we can have generic text in the modal and its resolutions, without the need to include the target module/colonist's information.
+
+### 5. Push the random events to the game's database.
+
+### 6. Create an endpoint for retrieving random events and add it to the backend's index file.
+
+### 7. Create a frontend server function that can retrieve a random event from the server, and have the Engine call it in lieu of reading from the constants file.
 
 ## Chapter X: In-Game Notifications